+++ lib/Conversion/TritonGPUToLLVM/CMakeLists.txt
@@ -2,12 +2,12 @@ add_mlir_conversion_library(TritonGPUToLLVM
     TritonGPUToLLVM.cpp
     TritonGPUToLLVMPass.cpp
     PTXAsmFormat.cpp
-    ConvertLayoutOpConversions.cpp
-    ElementwiseOpConversions.cpp
-    ViewOpConversions.cpp
-    LoadStoreOpConversions.cpp
-    DotOpConversions.cpp
-    ReduceOpConversions.cpp
+    ConvertLayoutOpToLLVM.cpp
+    ElementwiseOpToLLVM.cpp
+    ViewOpToLLVM.cpp
+    LoadStoreOpToLLVM.cpp
+    DotOpToLLVM.cpp
+    ReduceOpToLLVM.cpp
 
     ADDITIONAL_HEADER_DIRS
     ${PROJECT_SOURCE_DIR}/include/triton/Conversion/TritonGPUToLLVM
