+++ lib/Conversion/TritonGPUToLLVM/ConvertLayoutOpToLLVM.cpp
@@ -1,4 +1,4 @@
-#include "ConvertLayoutOpConversions.h"
+#include "ConvertLayoutToLLVM.h"
 #include "DotOpHelpers.h"
 
 using ::mlir::LLVM::DotOpFMAConversionHelper;
