+++ lib/Conversion/TritonGPUToLLVM/DotOpToLLVM.cpp
@@ -1,4 +1,4 @@
-#include "DotOpConversions.h"
+#include "DotOpToLLVM.h"
 #include "DotOpHelpers.h"
 #include "Utility.h"
