+++ lib/Conversion/TritonGPUToLLVM/ElementwiseOpToLLVM.cpp
@@ -1,4 +1,4 @@
-#include "ElementwiseOpConversions.h"
+#include "ElementwiseOpToLLVM.h"
 
 using namespace mlir;
 using namespace mlir::triton;
