+++ lib/Conversion/TritonGPUToLLVM/ViewOpToLLVM.cpp
@@ -1,4 +1,4 @@
-#include "ViewOpConversions.h"
+#include "ViewOpToLLVM.h"
 #include "DotOpHelpers.h"
 
 using namespace mlir;
